+++ test/quack_db/integration/ecto/analytics_test.exs
@@ -73,6 +73,8 @@ defmodule QuackDB.Integration.Ecto.AnalyticsTest do
       [~s({"name":"salmon","kind":"fish","score":5}), ~N[2024-01-03 04:05:06]]
     ])
 
+    interval = Duration.new!(minute: 15)
+
     query =
       from(event in table,
         where: json_extract_string(event.payload, "$.kind") == "bird",
@@ -80,7 +82,8 @@ defmodule QuackDB.Integration.Ecto.AnalyticsTest do
           name: json_extract_string(event.payload, "$.name"),
           score: json_extract(event.payload, "$.score"),
           day: date_trunc("day", event.occurred_at),
-          bucket: time_bucket("1 day", event.occurred_at)
+          bucket: time_bucket("1 day", event.occurred_at),
+          fifteen_minute_bucket: time_bucket(^interval, event.occurred_at)
         }
       )
 
@@ -89,7 +92,8 @@ defmodule QuackDB.Integration.Ecto.AnalyticsTest do
                name: "duck",
                score: "10",
                day: ~N[2024-01-02 00:00:00.000000],
-               bucket: ~N[2024-01-02 00:00:00.000000]
+               bucket: ~N[2024-01-02 00:00:00.000000],
+               fifteen_minute_bucket: ~N[2024-01-02 03:00:00.000000]
              }
            ] = QuackDB.IntegrationRepo.all(query)
   end